Cleaned up helpers.
