+++ util/inBrowser.js
@@ -1,0 +1,7 @@
+'use strict';
+
+// Note: in iron-node window is defined - but it has window.process
+// which is not there in a real browser env
+var inBrowser = ( typeof window !== 'undefined' && !window.process );
+
+module.exports = inBrowser;
